requirements.txt
python-dotenv==1.0.0
slack-bolt==1.18.0
langchain==0.0.292
chromadb==0.4.18
sentence-transformers==2.2.2
rasa==3.6.6
activepieces-sdk==0.1.0
pydantic==2.3.0
numpy==1.24.3
pandas==2.1.0
pytest==7.4.2
